Move _move_tensors_to_cpu to util/tensor.py and fold its tests into test_multi_process.py
The helper is a tensor transform with no model state, so it sits next to
batch_to_device rather than in base/model.py, and all four families already import
from sentence_transformers.util. BaseModel keeps _report_worker_failure, which owns
the worker result protocol rather than a tensor fact.

tests/base/test_multi_process_worker.py rebuilt the fake queues that
tests/base/test_multi_process.py already provides, and its queue.Empty loop break
stopped working once the workers dropped that branch, so its two tests move there.
