+++ sentence_transformers/util/__init__.py
@@ -55,6 +55,7 @@ from .tensor import (
     _convert_to_batch,
     _convert_to_batch_tensor,
     _convert_to_tensor,
+    _move_tensors_to_cpu,
     batch_to_device,
     cat_padded_token_embeddings,
     compute_count_vector,
@@ -130,6 +131,7 @@ __all__ = [
     "_convert_to_batch",
     "_convert_to_batch_tensor",
     "_convert_to_tensor",
+    "_move_tensors_to_cpu",
     "batch_to_device",
     "cat_padded_token_embeddings",
     "compute_count_vector",
